Introduced logging into the lambda and modified tests.
The lambda now logs the response code, body and x-message-id from the header. The tests were modified to accommodate these changes.

The tests were also modified to test the argument of the sendgrid send function instead of the output of the lambda.

Carrot symbols were also removed from versions in package.json.